requirements.txt
asyncpg
python-dotenv
httpx
beautifulsoup4